+++ qt/scientific_interfaces/Indirect/FunctionTemplateBrowser.h
@@ -68,7 +68,7 @@ public:
   virtual void setResolution(std::string const &name,
                              TableDatasetIndex const &index) = 0;
   virtual void setResolution(
-      const std::vector<std::pair<std::string, int>> &fitResolutions) = 0;
+      const std::vector<std::pair<std::string, size_t>> &fitResolutions) = 0;
   virtual void setQValues(const std::vector<double> &qValues) = 0;
 
 signals:
